Prevent post-install aliasing to occur in dev environment

+++ scripts/postinstall.js
@@ -20,20 +20,24 @@ if (fs.existsSync(targetFilePath)) {
 }
 
 /* OUI -> EUI Aliases */
-const shell = require('shelljs');
+const { INIT_CWD, PWD = process.cwd() } = process.env;
+// Only run when installed as a dep
+if (!INIT_CWD?.startsWith?.(PWD)) {
+  const shell = require('shelljs');
 // Clean up and recreate the folders
-shell.rm('-rf', 'src/themes/eui');
-shell.rm('-rf', 'src/themes/eui-amsterdam');
-shell.mkdir('-p', 'src/themes/eui');
-shell.mkdir('-p', 'src/themes/eui-amsterdam');
+  shell.rm('-rf', 'src/themes/eui');
+  shell.rm('-rf', 'src/themes/eui-amsterdam');
+  shell.mkdir('-p', 'src/themes/eui');
+  shell.mkdir('-p', 'src/themes/eui-amsterdam');
 
-shell.cp('-fR', 'src/themes/oui/oui_colors_dark.scss', 'src/themes/eui/eui_colors_dark.scss');
-shell.cp('-fR', 'src/themes/oui/oui_colors_light.scss', 'src/themes/eui/eui_colors_light.scss');
-shell.cp('-fR', 'src/themes/oui/oui_globals.scss', 'src/themes/eui/eui_globals.scss');
+  shell.cp('-fR', 'src/themes/oui/oui_colors_dark.scss', 'src/themes/eui/eui_colors_dark.scss');
+  shell.cp('-fR', 'src/themes/oui/oui_colors_light.scss', 'src/themes/eui/eui_colors_light.scss');
+  shell.cp('-fR', 'src/themes/oui/oui_globals.scss', 'src/themes/eui/eui_globals.scss');
 
-shell.cp('-fR', 'src/themes/oui-cascadia/global_styling', 'src/themes/eui-amsterdam/global_styling');
-shell.cp('-fR', 'src/themes/oui-cascadia/overrides', 'src/themes/eui-amsterdam/overrides');
-shell.cp('-fR', 'src/themes/oui-cascadia/oui_cascadia_colors_dark.scss', 'src/themes/eui-amsterdam/eui_amsterdam_colors_dark.scss');
-shell.cp('-fR', 'src/themes/oui-cascadia/oui_cascadia_colors_light.scss', 'src/themes/eui-amsterdam/eui_amsterdam_colors_light.scss');
-shell.cp('-fR', 'src/themes/oui-cascadia/oui_cascadia_globals.scss', 'src/themes/eui-amsterdam/eui_amsterdam_globals.scss');
+  shell.cp('-fR', 'src/themes/oui-cascadia/global_styling', 'src/themes/eui-amsterdam/global_styling');
+  shell.cp('-fR', 'src/themes/oui-cascadia/overrides', 'src/themes/eui-amsterdam/overrides');
+  shell.cp('-fR', 'src/themes/oui-cascadia/oui_cascadia_colors_dark.scss', 'src/themes/eui-amsterdam/eui_amsterdam_colors_dark.scss');
+  shell.cp('-fR', 'src/themes/oui-cascadia/oui_cascadia_colors_light.scss', 'src/themes/eui-amsterdam/eui_amsterdam_colors_light.scss');
+  shell.cp('-fR', 'src/themes/oui-cascadia/oui_cascadia_globals.scss', 'src/themes/eui-amsterdam/eui_amsterdam_globals.scss');
+}
 /* End of Aliases */
